Fix bug in Chatpage
